refactor(website): move apiReference module lookup into domain
`modulesBySlug` and `resolveModule` sat in `page/apiReference/index.ts`
alongside its re-exports. The barrel already re-exports `domain.ts` with
`export *`, so moving them there leaves `resolveModule` resolving the same
way and none of the three call sites change.

`modulesBySlug` builds the record with `Record.fromIterableBy` instead of
mapping to tuples and folding them with `Record.fromEntries`.

Extracted from #926.
